+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+* rtm/delete delete a stream, instead of writing a null entry in it
+* health check delete temp channel/streams + test read/write/delete
 * Add optional http test in health-check, with --http_extra_check
 
 ## [1.6.11] - 2019-08-28